Move FTP client + Command Manager to other files + add code doc
